+++ src/extension.ts
@@ -1,7 +1,7 @@
 // The module 'vscode' contains the VS Code extensibility API
 // Import the module and reference it with the alias vscode in your code below
 import * as vscode from 'vscode';
-import { extractCosmosDiagnostics } from './core';
+import { extractCosmosDiagnostics, extractInCompleteCosmosDiagnostics } from './core';
 
 // This method is called when your extension is activated
 // Your extension is activated the very first time the command is executed
@@ -9,17 +9,23 @@ export function activate(context: vscode.ExtensionContext) {
 
 	// Use the console to output diagnostic information (console.log) and errors (console.error)
 	// This line of code will only be executed once when your extension is activated
-	console.log('Congratulations, your extension "cosmosdiagnosticsextractor" is now active!');
+	console.log('Congratulations, your extension "cosmosdiagnosticsformatter" is now active!');
 
 	// The command has been defined in the package.json file
 	// Now provide the implementation of the command with registerCommand
 	// The commandId parameter must match the command field in package.json
-	const disposable = vscode.commands.registerCommand('cosmosdiagnosticsextractor.extractCosmosDiagnostics', () => {
+	const disposable1 = vscode.commands.registerCommand('cosmosdiagnosticsformatter.extractCosmosDiagnostics', () => {
 		// The code you place here will be executed every time your command is executed
 		extractCosmosDiagnostics();
 	});
+	context.subscriptions.push(disposable1);
 
-	context.subscriptions.push(disposable);
+	// Register the command to extract InComplete Cosmos diagnostics
+	const disposable2 = vscode.commands.registerCommand('cosmosdiagnosticsformatter.extractInCompleteCosmosDiagnostics', () => {
+		// The code you place here will be executed
+		extractInCompleteCosmosDiagnostics();
+	});
+	context.subscriptions.push(disposable2);
 }
 
 // This method is called when your extension is deactivated