+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Change Log
 
-All notable changes to the "cosmosdiagnosticsextractor" extension will be documented in this file.
-
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+## 0.0.2
+- Update title of exiting command .
+- Add another command "Cosmos Diagnostics: Extract, Fix and Format" to handle incomplete logs.
 
 ## 0.0.1